Split TXT record kv pair and propagate BString type value
Following TXT Record format for DNS-SD detailed in RFC 6763,
an attribute contains a key, value pairing where value is
an opaque binary data. The spec also differentiates
no-value attribute (key) from empty-value attribute (key=).
The keys are mandated to be unique in a service type context,
and duplicates in a single TXT record are suggested
to be silently ignored.

As such, this commit makes the following changes:
- Split the TXT record into key value pairs
- Use HashMap instead of Vector of TXT records
- Define an explicit enum for TXT record value type
- Propagate BString value, avoiding lossy translation to UTF-8 string