Implement multihome mDNS with a single UDP socket.
Utilizes our fork of the `multicast-socket` crate which was
needed to add a proper `tokio` API, and the new AsyncUdpSocket
trait implemented in the previous commit.

Multi-homing is currently behind a feature flag.

Also fix a pre-existing bug in the crate with the interval semantics;
this crate was originally `async-std` only and we added the Tokio impl.
Apparently, with `async-std` the interval would fire immediately and
then every interval after, but with `tokio` it doesn't. So we weren't
actually sending a query when starting the listener.